api: separate tags.dart into multiple files
