+++ lib/api/tags/index.dart
@@ -1,0 +1,11 @@
+library tag_manager;
+
+import 'dart:convert';
+import 'dart:io';
+
+import 'package:http/http.dart' as http;
+import 'package:path/path.dart' as p;
+
+part 'autotag.dart';
+part 'identification.dart';
+part 'match.dart';
